+++ app/scripts/migrations/index.js
@@ -30,4 +30,5 @@ module.exports = [
   require('./017'),
   require('./018'),
   require('./019'),
+  require('./020'),
 ]